requirements.txt
accelerate==0.27.2
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.0
aliyun-python-sdk-kms==2.16.2
annotated-types==0.6.0
anyio==4.2.0
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
autofaiss==2.17.0
Babel==2.14.0
bce-python-sdk==0.9.5
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.3
certifi 
cffi==1.16.0
charset-normalizer 
click==8.1.7
cmake==3.28.3
contourpy==1.2.0
crcmod==1.7
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.9
dataclasses-json==0.6.4
datasets==2.18.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
einops==0.7.0
embedding-reader==1.7.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
fastapi==0.110.0
filelock 
fire==0.5.0
Flask==3.0.2
flask-babel==4.0.0
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.5.4
greenlet==3.0.3
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.21.4
idna 
imageio==2.34.0
imgaug==0.4.0
importlib_metadata==7.0.2
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
jieba==0.42.1
Jinja2 
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==17.0.6
llama-index==0.10.19
llama-index-agent-openai==0.1.5
llama-index-cli==0.1.9
llama-index-core==0.10.19
llama-index-embeddings-openai==0.1.6
llama-index-indices-managed-llama-cloud==0.1.4
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.9
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.9
llama-index-readers-llama-parse==0.1.3
llama-parse==0.3.8
llamaindex-py-client==0.1.13
llvmlite==0.42.0
lmdb==1.4.1
lmdeploy==0.2.5
loguru==0.7.2
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
mmengine-lite==0.10.3
modelscope==1.13.0
more-itertools==10.2.0
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx
nltk==3.8.1
numba==0.59.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-runtime-cu11==11.8.89
nvidia-nccl-cu11==2.20.5
openai==1.13.3
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
oss2==2.18.4
packaging==24.0
paddle-bfloat==0.1.7
paddlepaddle-gpu==2.6.0
pandas==2.2.1
parso==0.8.3
pdf2docx==0.5.8
peft==0.8.2
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
premailer==3.10.0
prompt-toolkit==3.0.43
protobuf==3.20.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyclipper==1.3.0.post5
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
PyDocX==0.9.10
Pygments==2.17.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pynvml==11.5.0
pyparsing==3.1.2
pypdf==4.1.0
pysqlite3==0.5.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
rarfile==4.1
regex==2023.12.25
requests 
rich==13.7.1
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
sentencepiece==0.2.0
shapely==2.0.3
shortuuid==1.0.13
similarities==1.1.7
simplejson==3.19.2
six 
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.28
SQLite3Database==0.2.0
stack-data==0.6.3
starlette==0.36.3
starlette-session==0.4.3
striprtf==0.0.26
sympy
tenacity==8.2.3
termcolor==2.4.0
text2vec==1.2.9
threadpoolctl==3.3.0
tifffile==2024.2.12
tiktoken==0.6.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
triton==2.2.0
typing-inspect==0.9.0
typing_extensions
tzdata==2024.1
urllib3
uvicorn==0.28.0
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zhconv==1.4.3
zipp==3.17.0